Fixes
Fixed a typo in COBRAparser regarding hmdb and added checks for reactions that are null for gpr and subsystems, logging the respective index.
Added two variants to ModelField.
Fixed runModelPolisher.sh. It should now only backup the original template.
Added logging for collections not in MIRIAM registry to BIGGAnnotation.
Changed jsbml version to 1.2 release, as it should be identical to the current snapshot.

+++ scripts/runModelPolisher.sh
@@ -102,7 +102,7 @@ fi
 
 "${DIR}/ModelPolisher.sh" $ARGS 
 
-if [ -e "${DIR}/ModelPolisherTemplate.sh" ] 
+if [ -e "${DIR}/ModelPolisherTemplate.sh" ] && [ ! -e "${DIR}/ModelPolisherTemplate.bckp" ];
 then
   mv "${DIR}/ModelPolisherTemplate.sh" "${DIR}/ModelPolisherTemplate.bckp"
 fi
